Added working version of drone deployment workflow
