--- servicexd/test/sade/option1.yml
@@ -1,57 +1,0 @@
-services:
-  reserve_port:
-    type: "action"
-    command: "python reserve_ports.py 4 ports.config"
-    stdout_file: "reserve_ports.log"
-    stderr_file: "reserve_ports_error.log"
-    state:
-      log:
-        source: "reserve_ports.log"
-        states:
-          complete: "ports and their PIDs written to"
-          confirmed: "ports reserved confirmed"
-
-  program1:
-    type: "action"
-    command: "./program1.sh"
-    stdout_file: "program1.log"
-    stderr_file: "program1_error.log"
-    state:
-      log:
-        source: "program1.log"
-        states:
-          ready: "program is ready"
-          complete: "program is completed"
-      env_var:
-        source: "PROGRAM1_STATUS"
-        states:
-          ready: "ready"
-    depends_on:
-      program2:
-        type: "log"
-        state: "ready"
-
-  program2:
-    type: "service"
-    command: "./program2.sh"
-    stdout_file: "program2.log"
-    stderr_file: "program2_error.log"
-    state:''''''''n
-      log:
-        source: "program2.log"
-        states:
-          ready: "program is ready"
-          complete: "program is completed"
-      url:
-        source: "http://example.com/status"
-        states:
-          running: "running"
-    depends_on:
-      reserve_port:
-        type: "log"
-        state: "complete"
-
-output:
-  state_times: "state_times.json"
-stop_signal: "/Users/saiful/CCL/cctools/servicexd/test/stop.txt"
-max_run_time: 50
