+++ servicexd/test/sade/reserve_ports.py
@@ -3,14 +3,39 @@ import socket
 import os
 import sys
 
+
+def find_and_reserve_ports(prefixes, output_file, repeat, start_port):
+    reserved_ports = []
+    next_port = start_port
+
+    for r in range(repeat):
+        for prefix_index, prefix in enumerate(prefixes):
+            port = find_unused_port(next_port)
+
+            pid = 1  # Placeholder PID
+            # pid = bind_port_with_nc(port)
+            # todo: check if bind is successful before adding to reserved port
+
+            reserved_ports.append((prefix, port, pid, r))
+            next_port = port + 1
+
+    with open(output_file, 'w') as f:
+        for i, (prefix, port, pid, index) in enumerate(reserved_ports):
+            # print(f"drone_{index}_{prefix}_port={port}\ndrone_{index}_{prefix}_pid={pid}")
+            f.write(f"drone_{index}_{prefix}_port={port}\ndrone_{index}_{prefix}_pid={pid}\n")
+
+
 def find_unused_port(start_port):
     while True:
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
-            try:
-                sock.bind(("", start_port))
-                return start_port
-            except OSError:
-                start_port += 1
+            with socket.socket(socket.AF_INET, socket.SOCK_DGRAM) as usock:
+                try:
+                    sock.bind(("", start_port))
+                    usock.bind(("", start_port))
+                    return start_port
+                except OSError:
+                    start_port += 1
+
 
 def bind_port_with_nc(port):
     try:
@@ -21,36 +46,21 @@ def bind_port_with_nc(port):
         print(f"Error binding port {port}: {e}")
         return None
 
-def find_and_reserve_ports(count, output_file, repeat):
-    reserved_ports = []
-    start_port = 4560
-
-    for r in range(repeat):
-        for _ in range(count):
-            port = find_unused_port(start_port)
-            pid = 0
-            reserved_ports.append((port, pid, r))
-            start_port = port + 1
-
-    with open(output_file, 'w') as f:
-        for i, (port, pid, index) in enumerate(reserved_ports):
-            port_index = i % count + 1
-            f.write(
-                f"port_{port_index}_{index} = {port}\npid_port_{port_index}_{index} = {pid}\n")
 
 def main():
-    if len(sys.argv) < 3 or len(sys.argv) > 4:
-        print(
-            "Usage: python reserve_ports.py <count> <output_file> [<repeat>]")
+    if len(sys.argv) < 5:
+        print("Usage: python reserve_ports.py <output_file> <repeat> <start_port> <prefix>")
         sys.exit(1)
 
-    count = int(sys.argv[1])
-    output_file = sys.argv[2]
-    repeat = int(sys.argv[3]) if len(sys.argv) == 4 else 1
+    output_file = sys.argv[1]
+    repeat = int(sys.argv[2])
+    start_port = int(sys.argv[3])
+    prefixes = sys.argv[4].split(',')
 
-    find_and_reserve_ports(count, output_file, repeat)
-    print(
-        f"Reserved {count * repeat} ports and their PIDs written to {output_file}")
+    find_and_reserve_ports(prefixes, output_file, repeat, start_port)
+
+    print(f"Reserved {len(prefixes) * repeat} ports and their PIDs written to {output_file}")
+
 
 if __name__ == "__main__":
     main()
