+++ servicexd/test/sade/sade-wf.yml
@@ -1,7 +1,7 @@
 services:
   reserve_port:
     type: "action"
-    command: "python3 /home/user/servicexd/test/reserve_ports_with_name.py /home/user/Firmware/build/px4_sitl_default/etc/init.d-posix/ports.config 8 4600 simulator,gcs_local,gcs_remote,offboard_local,onboard_payload_local,onboard_gimbal_local,typhoon_offboard_local,gazebo_master"
+    command: "python3 /tmp/sade/reserve_ports.py /home/user/Firmware/build/px4_sitl_default/etc/init.d-posix/ports.config 8 4600 simulator,gcs_local,gcs_remote,offboard_local,onboard_payload_local,onboard_gimbal_local,typhoon_offboard_local,gazebo_master"
     log_file: "/tmp/log/reserve_ports.log"
     error_file: "/tmp/log/reserve_ports_error.log"
     state:
@@ -26,7 +26,7 @@ services:
 
   gazebo_server:
     type: "service"
-    command: "/tmp/start_gazebo_server.sh"
+    command: "/tmp/sade/start_gazebo_server.sh"
     log_file: "/tmp/log/gazebo_server.log"
     error_file: "/tmp/log/gazebo_server_error.log"
     state:
@@ -35,17 +35,35 @@ services:
     depends_on:
       chmod_port_config: "final"
 
-  px4_0:
+  px4_instance_0:
     type: "service"
-    command: "/tmp/start_px4.sh"
+    command: "/tmp/sade/start_px4_instance.sh 0"
     log_file: "/tmp/log/px4_0.log"
     error_file: "/tmp/log/px4_0_error.log"
     state:
       log:
+        waiting_for_simulator: "Waiting for simulator to accept connection"
         ready: "Startup script returned successfully"
     depends_on:
       gazebo_server: "ready"
 
+  spawn_model_0:
+    type: "service"
+    command: "/tmp/sade/spawn_model.sh 0"
+    log_file: "/tmp/log/spawn_model_0.log"
+    error_file: "/tmp/log/spawn_model_0_error.log"
+    depends_on:
+      px4_instance_0: "waiting_for_simulator"
+
+  pose_sender:
+    type: "service"
+    command: "/tmp/sade/start_pose_sender.sh"
+    log_file: "/tmp/log/pose_sender.log"
+    error_file: "/tmp/log/pose_sender_error.log"
+    depends_on:
+      px4_instance_0: "ready"
+
 output:
   state_times: "/tmp/log/state_times.json"
 stop_signal: "/tmp/log/stop.txt"
+
